Create code6
This code retrieves country information using the REST Countries API (https://restcountries.com/). When the user clicks the "Search" button, it fetches the data for the specified country, extracts the relevant information, and displays it on the page.

To use this code, you'll need to include the HTML code in your web page and place the JavaScript code in a <script> tag or an external JavaScript file. Make sure you have a working internet connection to fetch data from the API.